# Optimize SM2 nearly 5 times