--- a/gm/gmsm/sm2/.gitignore
+++ b/gm/gmsm/sm2/.gitignore
@@ -1,0 +1,5 @@
+ifile
+ofile
+*.pem
+*.cpu
+*.test