… gm/gmsm/sm2/_lib/test.s
… gm/gmsm/sm2/avx_amd64.s
… gm/gmsm/sm2/opt_design.md
… gm/gmsm/sm2/p256.go
… gm/gmsm/sm2/p256_test.go
… gm/gmsm/sm2/sm2.go
… gm/gmsm/sm2/sm2_test.go