## Description

We optimized the SM2 algorithm by nearly 80%. 

_第四届中国软件开源创新大赛-开源项目创新赛命题组-s4plus云惊队_

## Type of change

- [x] Algorithm optimization.

## Brief of your solution

* Loop unrolling. We manually unroll some loops in code, as the go compiler does not have any loop optimization.
* 64-bit integer. We noticed that the previous version of SM2 implementation used 32-bit integer to do calculations, so we changed some data type from 32-bit integer to 64-bit integer, In this way, we can save some computing overhead.
* Memory optimization. 
  * We try to reduce allocations in program to improve the performance.
  * Due to the escape analysis of go compiler, sometimes it is too radical to allocate some unnecessary variables to the heap,  we try to manually modify the references of some pointers to bypass escape analysis, according to this paper: [Escape from escape analysis of Golang](https://dl.acm.org/doi/10.1145/3377813.3381368)
* SIMD instructions. According to this paper, [Parallel Implementation of SM2 Elliptic Curve Cryptography on Intel Processors with AVX2](https://link.springer.com/chapter/10.1007/978-3-030-55304-3_11), we use AVX and AVX2 to accelerate SM2 algorithm.
* For more detailed information, please check our optimization designing documentation.

## How Has This Been Tested?

We use the "tests" and "benchmarks" in `sm2_test.go` to test the correctness and performance. Here are the test results.

### Correctness

```bash
(dby_env) dby@user-SYS-2049U-TR4:~/crypto-opt/opt-crypto/gm/gmsm/sm2$ go test . -v
=== RUN   TestBigIntBits
[18446744073709551615 18446744069414584320 18446744073709551615 18446744069414584319]
--- PASS: TestBigIntBits (0.00s)
=== RUN   TestP256
--- PASS: TestP256 (0.16s)
=== RUN   TestSm2
true
Cipher text = [4 255 126 228 143 69 247 112 15 211 100 234 112 248 58 6 188 218 12 13 130 203 84 137 144 98 231 42 67 91 236 28 122 8 140 169 184 128 253 149 53 16 220 209 27 156 188 148 170 219 248 154 170 156 41 94 110 20 205 25 154 16 34 64 4 119 110 239 30 25 30 111 151 188 224 61 163 40 32 19 172 107 25 205 31 238 89 188 85 254 194 76 223 171 253 54 134 53 239 161 172 169 235]
clear text = 123456
Verify ok
Verify ok
CheckSignature ok
CheckSignature ok
--- PASS: TestSm2 (0.01s)
PASS
ok      github.com/xuperchain/crypto/gm/gmsm/sm2        0.173s
```

### Performance

**Testing environment**:

```
go version: go version go1.16.3 linux/amd64
cpu: cpu: Intel(R) Xeon(R) Gold 6130 CPU @ 2.10GHz
os: Linux user-SYS-2049U-TR4 4.15.0-151-generic #157~16.04.1-Ubuntu SMP Wed Jul 14 11:22:28 UTC 2021 x86_64 x86_64 x86_64 GNU/Linux
```

**The previous version**:

```bash
(dby_env) dby@user-SYS-2049U-TR4:~/crypto-opt/origin-crypto/gm/gmsm/sm2$ go test -run=Benchmark -bench=BenchmarkSM2$
goos: linux
goarch: amd64
pkg: github.com/xuperchain/crypto/gm/gmsm/sm2
cpu: Intel(R) Xeon(R) Gold 6130 CPU @ 2.10GHz
BenchmarkSM2-64              292           3846587 ns/op           82726 B/op       1748 allocs/op
PASS
ok      github.com/xuperchain/crypto/gm/gmsm/sm2        1.556s
```

**After optimization**:
```bash
(dby_env) dby@user-SYS-2049U-TR4:~/crypto-opt/opt-crypto/gm/gmsm/sm2$ go test -run=Benchmark -bench=BenchmarkSM2$
goos: linux
goarch: amd64
pkg: github.com/xuperchain/crypto/gm/gmsm/sm2
cpu: Intel(R) Xeon(R) Gold 6130 CPU @ 2.10GHz
BenchmarkSM2-64             1245            816125 ns/op           15222 B/op        280 allocs/op
PASS
ok      github.com/xuperchain/crypto/gm/gmsm/sm2        1.130s
```