@hongyanwang 2021-11-25
@BoyaoDing sm2中删除了原有的Sign方法，引用到这个方法的地方也要改下哦，可以关注下crypto/gm/sign/ecdsa.go